Fix 'done' flag always being true in API call.
API calls to write in resumable mode had the 'done' flag always set to
True. This was causing massive perforance issues, and chunk transfer
speed deteriorated as the file got transfered. I assume because the WD
software was having to perform extra operations / memory management to
reopen the file once it got told to append more data to a file that was
already closed.